Refactor api service methods tests using token.
