+++ fileHub/js-app/tests/app/services/state-management/download-file/tests-download-file-api-service.js
@@ -1,5 +1,6 @@
 import fetchMock from '../../../../../node_modules/fetch-mock/esm/client.js';
 import {ApiService} from '../../../../../app/services/api-service/api-service.js';
+import {getWindowMock} from '../../../mock-window-api-service.js';
 
 const {module, test} = QUnit;
 
@@ -11,8 +12,11 @@ export default () => module('downloadFile', () => {
       url: '/file/5',
       method: 'GET',
     }, {id});
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
+
     const result = await apiService.downloadFile( '5');
+
     assert.ok(fetch.called(), 'Should send a request');
     assert.equal(result.id, id, 'Should get id of uploaded file');
   });
@@ -24,7 +28,8 @@ export default () => module('downloadFile', () => {
       url: '/file/5',
       method: 'GET',
     }, 400);
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
 
     try {
       await apiService.downloadFile( '5');
@@ -42,7 +47,8 @@ export default () => module('downloadFile', () => {
       url: '/file/5',
       method: 'GET',
     }, 500);
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
 
     try {
       await apiService.downloadFile( '5');
