+++ fileHub/js-app/tests/app/services/state-management/fetch-current-directory/tests-fetch-current-folder-api-service.js
@@ -1,5 +1,6 @@
 import fetchMock from '../../../../../node_modules/fetch-mock/esm/client.js';
 import {ApiService} from '../../../../../app/services/api-service/api-service.js';
+import {getWindowMock} from '../../../mock-window-api-service.js';
 
 const {module, test} = QUnit;
 
@@ -12,7 +13,8 @@ export default () => module('getFolder', () => {
       url: '/folder/5',
       method: 'GET',
     }, {folder});
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
     const result = await apiService.getFolder('5');
 
     assert.ok(fetch.called(), 'Should send a request');
@@ -26,7 +28,8 @@ export default () => module('getFolder', () => {
       url: '/folder/5',
       method: 'GET',
     }, 400);
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
 
     try {
       await apiService.getFolder('5');
@@ -44,7 +47,8 @@ export default () => module('getFolder', () => {
       url: '/folder/5',
       method: 'GET',
     }, 500);
-    const apiService = new ApiService({fetch});
+    const mockWindow = getWindowMock(fetch);
+    const apiService = new ApiService(mockWindow);
 
     try {
       await apiService.getFolder('5');
